+++ modal/cloud_bucket_mount.py
@@ -16,7 +16,8 @@ class _CloudBucketMount:
 
     S3 buckets are mounted using [AWS S3 Mountpoint](https://github.com/awslabs/mountpoint-s3).
     S3 mounts are optimized for reading large files sequentially. It does not support every file operation; consult
-    [the AWS S3 Mountpoint documentation](https://github.com/awslabs/mountpoint-s3/blob/main/doc/SEMANTICS.md) for more information.
+    [the AWS S3 Mountpoint documentation](https://github.com/awslabs/mountpoint-s3/blob/main/doc/SEMANTICS.md)
+    for more information.
 
     **AWS S3 Usage**
 
@@ -25,8 +26,13 @@ class _CloudBucketMount:
 
     app = modal.App()  # Note: "app" was called "stub" up until April 2024
     secret = modal.Secret.from_dict({
+        # Required: specify the access key of your AWS account
         "AWS_ACCESS_KEY_ID": "...",
+        # Required: specify the secret access key of your AWS account
         "AWS_SECRET_ACCESS_KEY": "...",
+        # Optional: specify the region of your S3 bucket.
+        # This can help when automatic detection of the bucket region fails.
+        "AWS_REGION": "...",
     })
     @app.function(
         volumes={
@@ -43,8 +49,8 @@ class _CloudBucketMount:
 
     **Cloudflare R2 Usage**
 
-    Cloudflare R2 is [S3-compatible](https://developers.cloudflare.com/r2/api/s3/api/) so its setup looks very similar to S3.
-    But additionally the `bucket_endpoint_url` argument must be passed.
+    Cloudflare R2 is [S3-compatible](https://developers.cloudflare.com/r2/api/s3/api/) so its setup looks
+    very similar to S3. But additionally the `bucket_endpoint_url` argument must be passed.
 
     ```python
     import subprocess
@@ -71,8 +77,8 @@ class _CloudBucketMount:
     **Google GCS Usage**
 
     Google Cloud Storage (GCS) is partially [S3-compatible](https://cloud.google.com/storage/docs/interoperability).
-    Currently **only `read_only=True`** is supported for GCS buckets. GCS Buckets also require a secret with Google-specific
-    key names (see below) populated with a [HMAC key](https://cloud.google.com/storage/docs/authentication/managing-hmackeys#create).
+    GCS Buckets also require a secret with Google-specific key names (see below) populated with
+    a [HMAC key](https://cloud.google.com/storage/docs/authentication/managing-hmackeys#create).
 
     ```python
     import subprocess
@@ -88,7 +94,6 @@ class _CloudBucketMount:
                 bucket_name="my-gcs-bucket",
                 bucket_endpoint_url="https://storage.googleapis.com",
                 secret=gcp_hmac_secret,
-                read_only=True,  # writing to bucket currently unsupported
             )
         }
     )
@@ -100,6 +105,8 @@ class _CloudBucketMount:
     bucket_name: str
     # Endpoint URL is used to support Cloudflare R2 and Google Cloud Platform GCS.
     bucket_endpoint_url: Optional[str] = None
+
+    key_prefix: Optional[str] = None
 
     # Credentials used to access a cloud bucket.
     # If the bucket is private, the secret **must** contain AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.
@@ -122,13 +129,10 @@ def cloud_bucket_mounts_to_proto(mounts: List[Tuple[str, _CloudBucketMount]]) ->
                 bucket_type = api_pb2.CloudBucketMount.BucketType.R2
             elif parse_result.hostname.endswith("storage.googleapis.com"):
                 bucket_type = api_pb2.CloudBucketMount.BucketType.GCP
-                if not mount.read_only:
-                    raise ValueError(
-                        f"CloudBucketMount of '{mount.bucket_name}' is invalid. Writing to GCP buckets with modal.CloudBucketMount in currently unsupported."
-                    )
             else:
                 logger.warn(
-                    "CloudBucketMount received unrecognized bucket endpoint URL. Assuming AWS S3 configuration as fallback."
+                    "CloudBucketMount received unrecognized bucket endpoint URL. "
+                    "Assuming AWS S3 configuration as fallback."
                 )
                 bucket_type = api_pb2.CloudBucketMount.BucketType.S3
         else:
@@ -138,6 +142,11 @@ def cloud_bucket_mounts_to_proto(mounts: List[Tuple[str, _CloudBucketMount]]) ->
         if mount.requester_pays and not mount.secret:
             raise ValueError("Credentials required in order to use Requester Pays.")
 
+        if mount.key_prefix and not mount.key_prefix.endswith("/"):
+            raise ValueError("key_prefix will be prefixed to all object paths, so it must end in a '/'")
+        else:
+            key_prefix = mount.key_prefix
+
         cloud_bucket_mount = api_pb2.CloudBucketMount(
             bucket_name=mount.bucket_name,
             bucket_endpoint_url=mount.bucket_endpoint_url,
@@ -146,6 +155,7 @@ def cloud_bucket_mounts_to_proto(mounts: List[Tuple[str, _CloudBucketMount]]) ->
             read_only=mount.read_only,
             bucket_type=bucket_type,
             requester_pays=mount.requester_pays,
+            key_prefix=key_prefix,
         )
         cloud_bucket_mounts.append(cloud_bucket_mount)